Massive refactoring.
