+++ src/main/java/io/anserini/collection/CarCollection.java
@@ -24,6 +24,7 @@ import java.io.BufferedReader;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.HashSet;
@@ -71,7 +72,7 @@ public class CarCollection extends DocumentCollection<CarCollection.Document> {
 
     public Segment(BufferedReader bufferedReader) throws IOException {
       super(bufferedReader);
-      stream = new ReaderInputStream(bufferedReader);
+      stream = new ReaderInputStream(bufferedReader, StandardCharsets.UTF_8);
       iter = DeserializeData.iterableParagraphs(stream).iterator();
     }
