--- src/main/python/README.md
@@ -1,10 +1,0 @@
-### Requirements
-
-Python>=2.6 or Python>=3.5
-`pip install -r src/main/python/requirements.txt`
-
-### Run
-
-```
-python src/main/python/run_regression.py --collection disk12 --index
-```
