--- src/main/python/generate_regressions_script.py
@@ -1,31 +1,0 @@
-#
-# Anserini: A Lucene toolkit for reproducible information retrieval research
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-from os import listdir
-
-print('#!/bin/sh\n')
-print('# This script is automatically generated by running the following command:')
-print('#   python src/main/python/generate_regressions_script.py > bin/run_regressions.sh\n')
-
-for f in sorted(listdir('src/main/resources/regression')):
-    if not f.endswith('yaml'):
-        continue
-    # Skip the full cw12 regression since it takes too much space and takes too long.
-    if f.endswith('cw12.yaml'):
-        continue
-
-    r = f.split('.yaml')[0]
-    print(f'python src/main/python/run_regression.py --index --verify --search --regression {r} > logs/log.{r} 2>&1')
