--- src/main/python/requirements.txt
@@ -1,2 +1,0 @@
-matplotlib==2.2.2
-pyyaml>=4.2b1
