+++ src/main/resources/docgen/templates/beir-v1.0.0-arguana-splade-distil-cocodenser-medium.template
@@ -3,7 +3,8 @@
 **Model**: SPLADE-distil CoCodenser Medium
 
 This page describes regression experiments, integrated into Anserini's regression testing framework, using SPLADE-distil CoCodenser Medium on [BEIR (v1.0.0) &mdash; ArguAna](http://beir.ai/).
-The SPLADE-distil CoCodenser Medium model is open-sourced by [Naver Labs Europe](https://europe.naverlabs.com/research/machine-learning-and-optimization/splade-models).
+SPLADE-distil CoCodenser Medium is an intermediate version between [SPLADEv2](https://arxiv.org/abs/2109.10086) and [SPLADE++](https://arxiv.org/abs/2205.04733), where the model used distillation (as in SPLADEv2), but started with the CoCondenser pre-trained model.
+See the [official SPLADE repo](https://github.com/naver/splade) for more details.
 
 The exact configurations for these regressions are stored in [this YAML file](${yaml}).
 Note that this page is automatically generated from [this template](${template}) as part of Anserini's regression pipeline, so do not modify this page directly; modify the template instead and then run `bin/build.sh` to rebuild the documentation.
@@ -19,7 +20,6 @@ python src/main/python/run_regression.py --index --verify --search \
 
 We make available a version of the BEIR-v1.0.0 arguana corpus that has already been processed with SPLADE-distil CoCodenser Medium, i.e., gone through document expansion and term reweighting.
 Thus, no neural inference is involved.
-For details on how to train SPLADE-distil CoCodenser Medium and perform inference, please see [guide provided by Naver Labs Europe](https://github.com/naver/splade/tree/main/anserini_evaluation).
 
 Download the corpus and unpack into `collections/`:
