--- src/main/resources/featureExtractors/defaultTwitterExtractor.txt
@@ -1,28 +1,0 @@
-{
- extractors: [
-  {
-    name:MatchingTermCount
-  },
-  {
-    name: QueryLength
-  },
-  {
-    name:QueryLength
-  },
-  {
-    name: TwitterHashtagCount
-  },
-  {
-    name: IsTweetReply
-  },
-  {
-    name: TwitterLinkCount
-  },
-  {
-    name : TwitterFollowerCount
-  },
-  {
-    name: TwitterFriendCount
-  }
- ]
-}
