+++ src/test/java/io/anserini/index/GeoIndexerTestBase.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.anserini;
+package io.anserini.index;
 
 import io.anserini.index.Constants;
 import org.apache.lucene.document.Document;
